+++ openHABWatch Extension/Views/Rows/SegmentRow.swift
@@ -9,7 +9,7 @@
 //
 // SPDX-License-Identifier: EPL-2.0
 
-import OpenHABCoreWatch
+import OpenHABCore
 import os.log
 import SwiftUI
 
@@ -39,7 +39,7 @@ struct SegmentRow: View {
                     DetailTextLabelView(widget: widget)
                 }
                 Picker("Picker", selection: valueBinding) {
-                    ForEach(0 ..< widget.mappingsOrItemOptions.count) {
+                    ForEach(0 ..< widget.mappingsOrItemOptions.count, id: \.self) {
                         Text(self.widget.mappingsOrItemOptions[$0].label).tag($0)
                     }
                 }